Renames IndexLevelEngine->HierarchicalLocMap and moves to loc_map.py
